+++ src/astedit/draw2d.c
@@ -189,7 +189,7 @@ void draw_colored_rect(int x, int y, int w, int h,
 
 void draw_rgba_texture_rect(Texture texture,
         int x, int y, int w, int h,
-        float texX, float texY, float texW, float texH)
+        int texX, int texY, int texW, int texH)
 {
         static struct TextureVertex2d rp[6];
